NumberEnergy/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
SQLAlchemy==2.0.40
alembic==1.15.2
Mako==1.3.10
bcrypt==4.3.0
stripe==12.0.1
requests==2.32.3
urllib3==2.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
colorama==0.4.6
typing-extensions==4.13.2
